// src/shared/models/api-responses/initiateUserResponse.ts
import { Community } from "../models/community";
import { User } from "../models/member";

export interface ValidateUserData {
  accessToken: string;
  appAccess: boolean;
  community: Community;
  hasAnswers: boolean;
  refreshToken: string;
  user: User;
}
export interface ValidateUserResponse {
  success: boolean;
  data?: ValidateUserData;
  errorMessage?: string | any;
}
export interface InitiateUserData {
  accessToken: string;
  appAccess: boolean;
  community: Community;
  hasAnswers: boolean;
  refreshToken: string;
  user: User;
}
export interface InitiateUserResponse {
  success: boolean;
  data?: InitiateUserData;
  errorMessage?: string | any;
}